dtschema: error: Print only the basename of the file
In order to reduce the length of an error line and stay as much as possible
within the 80 chars limit, let's only print the basename instead of the
full path.

The path itself should be pretty obvious depending on the configuration,
and if we have subfolders, the likeliness of having two DT with the exact
same name is pretty low.

Signed-off-by: Maxime Ripard <maxime.ripard@bootlin.com>