schemas: root-node: Introduce 'chassis-type' property
This patch adds an optional 'chassis-type' property to the root node in
order to specify the device form factor, mimicking what is already
available through DMI on ACPI-based systems.

Such information can be useful (for example, in order to implement
different behavior on a desktop and on a smartphone) but is not
currently available on DT-based systems.

Possible values are identical to those used by systemd's hostnamed.

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>