Switch to jsonschema 4.18+
We've been stuck on jsonschema 4.17 because 4.18 changed in
non-compatible ways with how RefResolver worked. Upstream was not
willing to fix the issues claiming our usage was incorrect.

The RefResolver interface is deprecated anyways so it isn't really worth
the effort to fix or find a work-around when we need to move to the
Registry interface anyways.

In 4.17, the Draft2019-09 meta-schema was (silently) broken due to
'$recursiveRef' not being handled, so we were stuck using Draft7
validator in some places. Now those are all updated to Draft2019-09.

Setting '$schema' to our meta-schema seems to not be how extending
the meta-schema is supposed to be done. '$schema' is supposed to be
which dialect of json-schema we are using (Draft 2019-09) rather than
our extended meta-schema. Going against this seems to cause problems
with referencing the core schema within our schema, so that has been
dropped in favor of validating against both meta-schemas separately.
Further reworking of how our meta-schema is structured is needed.

4.18+ has been out for some time now and is what distros are packaging.
Also, 4.17 seems to be broken with python 3.14 and a fix by upstream
seems unlikely. So now is the time to switch.

Signed-off-by: Rob Herring <robh@kernel.org>