dt-bindings: u-boot: Add variables for bootscript location
Add bootscr-address and bootscr-ram-offset properties to help in easier
picking of boot script file when automated flows are used.

The bootscr-address holds the full 64 bit address of the bootscript
file.

The bootscr-ram-offset holds the offset address of the bootscript file from
the start of the ram base in systems with RAM detection.

Co-develop-by: Algapally Santosh Sagar <santoshsagar.algapally@amd.com>
Signed-off-by: Algapally Santosh Sagar <santoshsagar.algapally@amd.com>
Signed-off-by: Michal Simek <michal.simek@amd.com>