schemas: iio: add io-backends schema
IIO backends are devices connected to another IIO devices
(consumers of the backends) providing services so that the consumer IIO device
can be fully functional. One typical example of such an Hardware arrangement
would be an high speed ADC that is connected to an FPGA IP core (through a
serial data interface as LVDS) capable of handling the high sample rates.
Examples of the provided services would be channel enablement/disablement.

'#io-backend-cells' is also being added to the iio schema.

While at it, fixed a typo in the top level description 'thes -> they'.

Acked-by: Olivier Moysan <olivier.moysan@foss.st.com>
Acked-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Signed-off-by: Nuno Sa <nuno.sa@analog.com>