schemas: memory: describe the Qualcomm ddr_device_type property
Qualcomm bootloaders will patch the /memory@ node with the DDR memory
type property. The property name breaks defined standards and the
property is not compatible with the defined LPDDR schema. However given
that determining memory type is required by some drivers to properly
program the hardware and the bootloaders are deployed in a variety of
device which can not and will not be updated by the ODMs (because
significant amount of devices has reached EOL) and the users also can
not update the bootloader (thanks to the SecureBoot / bootloader
signatures), document this property to be used in this particular use
case only, not to be used manually etc.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>