schemas: pci: bridge: Document WAKE# interrupt properties
WAKE# sideband interrupt is used by the PCIe devices to signal the host to
re-establish power and reference clocks while waking from D3Cold/L2 state.

This is based on the DT bindings patch proposed in LKML:
https://lore.kernel.org/linux-pci/20230208111645.3863534-2-mmaddireddy@nvidia.com/

In that patch, there were 2 interrupts mentioned: "wake" and "pci", where the
latter one was described as "legacy PCI interrupt". But those legacy interrupts
are already defined as "INT-{A,B,C,D}" in pci-device.yaml. So I removed that one
and just kept "wake".

Most of the platforms route the WAKE# GPIO to the PCI bridges (slots) in
hardware. But some platforms like Qcom SoCs, do not do that and for those
platforms the WAKE# interrupt is described in the controller binding itself.

Signed-off-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>