schemas: wakeup-source: Possibility for power states
wakeup-source in its current form ignores power states. But with modern
SoCs there are many powerstates that have a different set of devices
enabled. Also there are more complex wakeup mechanisms available.

As an example TI's am62 SoC has a power state called 'Partial-IO' in
which most of the SoC is powered off and only a small device logic is
present for a few select devices that can still wakeup the SoC. But most
devices that are usually able to wakeup the SoC are not capable to do it
in Partial-IO.

In order to properly describe the wakeup-source property depending on
the power state, this patch extends the definition to be either boolean
or a list of power state names in which this device is capable of
wakeup.

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>