schemas: wakeup-source: Possibility for system states
wakeup-source in its current form ignores system states. But with modern
SoCs there are many different system states that have a different set of
devices enabled. Also there are more complex wakeup mechanisms
available.

As an example TI's am62 SoC has a power state called 'Partial-IO' in
which most of the SoC is powered off and only the pin logic of a few
pins is active that can wakeup the SoC. But most devices that are
usually able to wakeup the SoC are not capable to do it in Partial-IO.

In order to properly describe the wakeup-source property depending on
the system state, this patch extends the definition to be either boolean
or a list of system state phandles in which this device is capable of
wakeup.

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>