schemas: system-idle-state: Add system state definition
Add a new binding to describe system states. Many new SoCs have
a number of different states the system as a whole can be in.
Certain devices or parts of the SoC could be for example powered off.
As an example am62a and am62p have states called 'Partial-IO', 'IO+DDR'
and 'MCU Only' which serve different needs and have different parts
enabled/disabled.

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>