schemas: chosen: describe new arm,security-state property
We once again stumbled over errata fixups that were applied twice, first in
the barebox bootloader and then in the Linux kernel.
This went unnoticed at first, but with OP-TEE enabled as a secure
monitor, the system hangs very early before regular console output is
available.

The Linux kernel errata fixes for non-ARCH_MULTIPLATFORM will likely never
go away, because bootloaders may not be updatable in the field and sometimes
don't even suffer from the consequences, e.g. because they affect only SMP.

Therefore, let's add at least a way for the bootloader to tell the OS about
whether it's running in the secure world or not.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>