schemas: chosen: document devicetree-path property
With more and more laptops, dev boards, and other products shipping with
both an EFI capable bootloader and relying on devicetree, there is an
increasing problem with trying to ensure that a devicetree is loaded and
that it will work with the kernel/OS version being booted.

This is less of an issue on vertically integrated embedded platforms,
but with ARM SystemReady gaining prevalence and the abundance of cheap
dev boards there is a growing desire to boot generic OS installer/disk
images on these devices.

In some cases the platform doesn't see many devicetree changes and works
well with a range of kernel versions, the board can just provide it's
own devicetree in firmware and it will work just fine. However there are
many platforms where this is not the case, unintentional breaking
changes can be made to drivers or DT, or new features being enabled can
break compatibility with older kernels.

In these cases, the best way to maximise the chance that the generic OS
image will boot is if it can load the devicetree blob which was
installed alongside the kernel (something most mainstream distros do).

To enable this there has to be a consistent way to determine which dtb
(and overlays) were embedded into the firmware provided devicetree.

This new property enables exactly this usecase by making it possible to
embed the information into the dtb itself at build time.

A proposed implementation of this for U-Boot is available here [1].

[1]: https://lore.kernel.org/u-boot/20250610-b4-fdtfile-from-dt-v1-0-aec4df0fa53e@linaro.org/

Signed-off-by: Casey Connolly <casey.connolly@linaro.org>