schemas: pci: bridge: Remove unused 'supports-d3' property
This property was added to specify that the PCI bridge supports D3 state, as
it cannot be discovered by the OS using PMCSR registers. This was added to
use it with the PCI subsystem of Linux Kernel. But this property was never used,
as the PCI subsystem in Linux Kernel decided to assume that all PCI bridges of
the DT based platforms supports transitioning to D3 state by default.

So remove this unused property.

Signed-off-by: Manivannan Sadhasivam <manivannan.sadhasivam@oss.qualcomm.com>