dt-bindings: chosen: Add "power-state-change-reason" nvmem property
A lot of systems (especially embedded) store the last shutdown/reboot
reason in NVMEM. This could be anything from "over temperature" or
"brown-out" to an "EC panic" or watchdog reset.

There's currently no standard way for firmware to tell the kernel where
to find this information.

The /chosen node is already used for passing global, firmware-provided
data like 'bootargs' and 'stdout-path'. Let's use it for this, too.

This patch adds a binding for a new generic NVMEM cell name in /chosen:
"power-state-change-reason".

Example from the binding:
    chosen {
        nvmem-cells = <&pscr_cell>;
        nvmem-cell-names = "power-state-change-reason";
    };

Signed-off-by: Oleksij Rempel <o.rempel@pengutronix.de>