Add yamllint config and fix all errors
Add configuration for yamllint and fix all the errors/warnings.
The configuration matches what is used in the Linux kernel tree except
single quotes are required and 'check-keys' option is enabled.
'check-keys' depends on a newish version of yamllint, so it hasn't been
enabled in the Linux kernel tree yet.

Signed-off-by: Rob Herring (Arm) <robh@kernel.org>