schemas: pci-bus-common: Add VIO 1.8V supply for M.2 slots
From the PCI Express M.2 specification rev. 5.1:

    The VIO 1.8 V power source is provided to supply the Adapter I/O
    buffer circuitry operating at 1.8 V sideband signaling. This power
    source is required for Adapters that support 1.8 V sideband signaling.
    Platforms that support 1.8 V sideband signaling must provide this
    power source. Adapters may provide 1.8 V sideband signaling based
    on the detection of this power source.

Add a property for describing this supply, so that it can be referenced
and controlled along with the other PCIe slot power sources.

The property itself is named vpcie1v8-supply instead of vio1v8-supply
as the former name already exists in some device trees.

Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>