schemas: pci: bridge: Remove unused "supports-d3" property
This property was added to specify whether the bridge supports transitioning to
D3 state (D3Cold) or not as needed by Linux. But it turns out that this
information could be inferred from the existence of the supply properties in the
bridge node. So this property was never used at all. So remove it.

Signed-off-by: Manivannan Sadhasivam <manivannan.sadhasivam@oss.qualcomm.com>